+++ torchsynth/torchmodule.py
@@ -25,11 +25,7 @@ class TorchSynthModule(nn.Module):
     TODO: Later, we should deprecate SynthModule and fold everything into here.
     """
 
-    def __init__(
-            self,
-            sample_rate: int = SAMPLE_RATE,
-            buffer_size: int = BUFFER_SIZE
-    ):
+    def __init__(self, sample_rate: int = SAMPLE_RATE, buffer_size: int = BUFFER_SIZE):
         """
         NOTE:
         __init__ should only set parameters.
@@ -118,20 +114,138 @@ class TorchSynthModule(nn.Module):
         return self.torchparameters[parameter_id].from_0to1()
 
 
-class TorchADSR(TorchSynthModule):
+class TorchSynthModule1D(TorchSynthModule):
+    """
+    Base class for synthesis modules, in torch.
+    All parameters are assumed to be 1D tensors,
+    the dimension size being the batch size.
+
+    WARNING: For now, TorchSynthModules should be atomic and not
+    contain other SynthModules.
+
+    TODO: Later, we should deprecate SynthModule and fold everything
+    into here.
+    """
+
+    # TODO: have these already moved to cuda
+    def __init__(self, batch_size: T, sample_rate: T = T(SAMPLE_RATE), buffer_size: T = T(BUFFER_SIZE)):
+        """
+        NOTE:
+        __init__ should only set parameters.
+        We shouldn't be doing computations in __init__ because
+        the computations will change when the parameters change.
+
+        batch_size is the number of settings we are rendering at once.
+        """
+        nn.Module.__init__(self)
+        assert sample_rate.shape == torch.Size([])
+        assert buffer_size.shape == torch.Size([])
+        self.sample_rate = sample_rate
+        self.buffer_size = buffer_size
+        self.batch_size = batch_size
+        self.torchparameters: nn.ParameterDict = nn.ParameterDict()
+
+    def to_buffer_size(self, signal: T) -> T:
+        return util.fix_length(signal, self.buffer_size)
+        #return util.fix_length2D(signal, self.buffer_size)
+
+    def seconds_to_samples(self, seconds: T) -> T:
+        # Do we want this?
+        #assert seconds.ndim == 1
+        return torch.round(seconds * self.sample_rate).int()
+
+    def _forward(self, *args: Any, **kwargs: Any) -> T:  # pragma: no cover
+        """
+        Each TorchSynthModule should override this.
+        """
+        raise NotImplementedError("Derived classes must override this method")
+
+    def forward1D(self, *args: Any, **kwargs: Any) -> T:  # pragma: no cover
+        """
+        Wrapper for _forward that ensures a buffer_size length output.
+        TODO: Make this forward() after everything is 1D
+        """
+        return self.to_buffer_size(self._forward(*args, **kwargs))
+
+    def forward(self, *args: Any, **kwargs: Any) -> T:  # pragma: no cover
+        """
+        Wrapper for _forward that ensures a buffer_size length output.
+        """
+        return self.forward1D(*args, **kwargs)
+
+    def add_parameters(self, parameters: List[TorchParameter]):
+        """
+        Add parameters to this SynthModule's torch parameter dictionary.
+        """
+        for parameter in parameters:
+            assert parameter.parameter_name not in self.torchparameters
+            assert parameter.shape == (self.batch_size,)
+            self.torchparameters[parameter.parameter_name] = parameter
+
+    def get_parameter(self, parameter_id: str) -> TorchParameter:
+        """
+        Get a single TorchParameter for this module
+
+        Parameters
+        ----------
+        parameter_id (str)  :   Id of the parameter to return
+        """
+        return self.torchparameters[parameter_id]
+
+    def get_parameter_0to1(self, parameter_id: str) -> T:
+        """
+        Get the value of a parameter in the range of [0,1]
+
+        Parameters
+        ----------
+        parameter_id (str)  :   Id of the parameter to return the value for
+        """
+        return self.torchparameters[parameter_id].item()
+
+    def set_parameter(self, parameter_id: str, value: T):
+        """
+        Update a specific parameter value, ensuring that it is within a specified
+        range
+
+        Parameters
+        ----------
+        parameter_id (str)  : Id of the parameter to update
+        value (T)           : Value to update parameter with
+        """
+        self.torchparameters[parameter_id].to_0to1(value)
+
+    def set_parameter_0to1(self, parameter_id: str, value: T):
+        """
+        Update a specific parameter with a value in the range [0,1]
+
+        Parameters
+        ----------
+        parameter_id (str)  : Id of the parameter to update
+        value (T)           : Value to update parameter with
+        """
+        assert torch.all(0 <= value <= 1)
+        self.torchparameters[parameter_id].data = value
+
+    def p(self, parameter_id: str) -> T:
+        """
+        Convenience method for getting the parameter value.
+        """
+        return self.torchparameters[parameter_id].from_0to1()
+
+
+class TorchADSR(TorchSynthModule1D):
     """
     Envelope class for building a control rate ADSR signal
     """
 
     def __init__(
         self,
-        a: float = 0.25,
-        d: float = 0.25,
-        s: float = 0.5,
-        r: float = 0.5,
-        alpha: float = 3.0,
-        sample_rate: int = SAMPLE_RATE,
-        buffer_size: int = BUFFER_SIZE
+        a: T,
+        d: T,
+        s: T,
+        r: T,
+        alpha: T,
+        **kwargs: Dict[str, T]
     ):
         """
         Parameters
@@ -145,38 +259,38 @@ class TorchADSR(TorchSynthModule):
         alpha               :   envelope curve, >= 0. 1 is linear, >1 is
                                 exponential.
         """
-        super().__init__(sample_rate=sample_rate, buffer_size=buffer_size)
+        super().__init__(batch_size=a.shape[0], **kwargs)
         self.add_parameters(
             [
                 TorchParameter(
                     value=a,
                     parameter_name="attack",
-                    parameter_range=ParameterRange(0.0, 2.0, curve="log")
+                    parameter_range=ParameterRange(0.0, 2.0, curve="log"),
                 ),
                 TorchParameter(
                     value=d,
                     parameter_name="decay",
-                    parameter_range=ParameterRange(0.0, 2.0, curve="log")
+                    parameter_range=ParameterRange(0.0, 2.0, curve="log"),
                 ),
                 TorchParameter(
                     value=s,
                     parameter_name="sustain",
-                    parameter_range=ParameterRange(0.0, 1.0)
+                    parameter_range=ParameterRange(0.0, 1.0),
                 ),
                 TorchParameter(
                     value=r,
                     parameter_name="release",
-                    parameter_range=ParameterRange(0.0, 5.0, curve="log")
+                    parameter_range=ParameterRange(0.0, 5.0, curve="log"),
                 ),
                 TorchParameter(
                     value=alpha,
                     parameter_name="alpha",
-                    parameter_range=ParameterRange(0.1, 6.0)
-                )
+                    parameter_range=ParameterRange(0.1, 6.0),
+                ),
             ]
         )
 
-    def _forward(self, note_on_duration: T = T(0)) -> T:
+    def _forward(self, note_on_duration: T) -> T:
         """Generate an ADSR envelope.
 
         By default, this envelope reacts as if it was triggered with midi, for
@@ -196,14 +310,15 @@ class TorchADSR(TorchSynthModule):
         If this is confusing, don't worry about it. ADSR's do a lot of work
         behind the scenes to make the playing experience feel natural.
         """
-
-        assert note_on_duration > 0
+        assert note_on_duration.ndim == 1
+        assert torch.all(note_on_duration > 0)
         num_samples = self.seconds_to_samples(note_on_duration)
 
         # Release decays from the last value of the attack-and-decay sections.
         ADS = self.note_on(num_samples)
         R = self.note_off(ADS[-1])
 
+        # Slightly worried this won't be performant on GPU?
         return torch.cat((ADS, R))
 
     def _ramp(self, duration: T, inverse: bool = False):
@@ -226,10 +341,11 @@ class TorchADSR(TorchSynthModule):
 
         """
 
-        assert duration.ndim == 0
+        assert duration.ndim == 1
+        # Unfortunately this won't work with 1D tensors
+        # and we can't have ramps in different sizes
         t = torch.arange(
-            self.seconds_to_samples(duration).item(),
-            device=duration.device
+            self.seconds_to_samples(duration), device=duration.device
         )
         ramp = t * (1 / duration) / self.sample_rate
 
@@ -253,11 +369,11 @@ class TorchADSR(TorchSynthModule):
         # `r`-length reverse ramp, reversed to descend to 0.
         return self._ramp(self.p("release"), inverse=True)
 
-    def note_on(self, num_samples):
-        assert self.attack.ndim == 1
-        assert self.decay.ndim == 1
+    def note_on(self, num_samples: T):
+        assert num_samples.ndim == 1
         out_ = torch.cat((self.attack, self.decay), 0)
 
+        # Slightly worried this won't be performant on GPU?
         # Truncate or extend based on sustain duration.
         if num_samples <= len(out_):
             out_ = out_[:num_samples]
@@ -269,7 +385,7 @@ class TorchADSR(TorchSynthModule):
 
         return out_
 
-    def note_off(self, last_val):
+    def note_off(self, last_val: T):
         return self.release * last_val
 
     def __str__(self):
@@ -311,25 +427,23 @@ class TorchVCO(TorchSynthModule):
         mod_depth: float = 50,
         phase: float = 0,
         sample_rate: int = SAMPLE_RATE,
-        buffer_size: int = BUFFER_SIZE
+        buffer_size: int = BUFFER_SIZE,
     ):
         TorchSynthModule.__init__(
-            self,
-            sample_rate=sample_rate,
-            buffer_size=buffer_size
+            self, sample_rate=sample_rate, buffer_size=buffer_size
         )
         self.add_parameters(
             [
                 TorchParameter(
                     value=midi_f0,
                     parameter_name="pitch",
-                    parameter_range=ParameterRange(0.0, 127.0)
+                    parameter_range=ParameterRange(0.0, 127.0),
                 ),
                 TorchParameter(
                     value=mod_depth,
                     parameter_name="mod_depth",
-                    parameter_range=ParameterRange(0.0, 127.0)
-                )
+                    parameter_range=ParameterRange(0.0, 127.0),
+                ),
             ]
         )
         # TODO: Make this a parameter too?
@@ -400,7 +514,7 @@ class TorchSineVCO(TorchVCO):
         midi_f0: float = 10.0,
         mod_depth: float = 50.0,
         phase: float = 0.0,
-        **kwargs
+        **kwargs,
     ):
         super().__init__(midi_f0=midi_f0, mod_depth=mod_depth, phase=phase, **kwargs)
 
@@ -422,10 +536,8 @@ class TorchFmVCO(TorchVCO):
     """
 
     def __init__(
-            self,
-            midi_f0: float = 10.0,
-            mod_depth: float = 50.0,
-            phase: float = 0.0):
+        self, midi_f0: float = 10.0, mod_depth: float = 50.0, phase: float = 0.0
+    ):
         super().__init__(midi_f0=midi_f0, mod_depth=mod_depth, phase=phase)
 
     def make_control_as_frequency(self, mod_signal: T):
@@ -457,14 +569,14 @@ class TorchSquareSawVCO(TorchVCO):
         midi_f0: float = 10.0,
         mod_depth: float = 50.0,
         phase: float = 0.0,
-     ):
+    ):
         super().__init__(midi_f0=midi_f0, mod_depth=mod_depth, phase=phase)
         self.add_parameters(
             [
                 TorchParameter(
                     value=shape,
                     parameter_name="shape",
-                    parameter_range=ParameterRange(0.0, 1.0)
+                    parameter_range=ParameterRange(0.0, 1.0),
                 )
             ]
         )
@@ -492,11 +604,7 @@ class TorchVCA(TorchSynthModule):
     Voltage controlled amplifier.
     """
 
-    def __init__(
-            self,
-            sample_rate: int = SAMPLE_RATE,
-            buffer_size: int = BUFFER_SIZE
-    ):
+    def __init__(self, sample_rate: int = SAMPLE_RATE, buffer_size: int = BUFFER_SIZE):
         super().__init__(sample_rate=sample_rate, buffer_size=buffer_size)
 
     def _forward(self, control_in: T, audio_in: T) -> T:
@@ -519,18 +627,14 @@ class TorchNoise(TorchSynthModule):
     **kwargs: see TorchSynthModule
     """
 
-    def __init__(
-            self,
-            ratio: float = 0.25,
-            **kwargs
-    ):
+    def __init__(self, ratio: float = 0.25, **kwargs):
         super().__init__(**kwargs)
         self.add_parameters(
             [
                 TorchParameter(
                     value=ratio,
                     parameter_name="ratio",
-                    parameter_range=ParameterRange(0.0, 1.0)
+                    parameter_range=ParameterRange(0.0, 1.0),
                 )
             ]
         )
@@ -549,11 +653,7 @@ class TorchSynthParameters(TorchSynthModule):
     A SynthModule that is strictly for managing parameters
     """
 
-    def __init__(
-            self,
-            sample_rate: int = SAMPLE_RATE,
-            buffer_size: int = BUFFER_SIZE
-    ):
+    def __init__(self, sample_rate: int = SAMPLE_RATE, buffer_size: int = BUFFER_SIZE):
         super().__init__(sample_rate, buffer_size)
 
     def _forward(self, *args: Any, **kwargs: Any) -> T:
@@ -571,11 +671,7 @@ class TorchSynth(nn.Module):
     buffer_size (int): number of samples expected at output of child modules
     """
 
-    def __init__(
-            self,
-            sample_rate: int = SAMPLE_RATE,
-            buffer_size: int = BUFFER_SIZE
-    ):
+    def __init__(self, sample_rate: int = SAMPLE_RATE, buffer_size: int = BUFFER_SIZE):
         super().__init__()
         self.sample_rate = T(sample_rate)
         self.buffer_size = T(buffer_size)
@@ -613,384 +709,63 @@ class TorchSynth(nn.Module):
             parameter.data = torch.rand_like(parameter)
 
 
-class TorchDrum(TorchSynth):
-    """
-    A package of modules that makes one drum hit.
-    """
-
-    def __init__(
-        self,
-        note_on_duration: float,
-        vco_ratio: float = 0.5,
-        pitch_adsr: TorchADSR = TorchADSR(),
-        amp_adsr: TorchADSR = TorchADSR(),
-        vco_1: TorchVCO = TorchSineVCO(),
-        vco_2: TorchVCO = TorchSquareSawVCO(),
-        noise: TorchNoise = TorchNoise(),
-        vca: TorchVCA = TorchVCA(),
-        **kwargs
-    ):
-        super().__init__(**kwargs)
-        assert note_on_duration >= 0
-
-        # We assume that sustain duration is a hyper-parameter,
-        # with the mindset that if you are trying to learn to
-        # synthesize a sound, you won't be adjusting the note_on_duration.
-        # However, this is easily changed if desired.
-        self.note_on_duration = T(note_on_duration)
-
-        # Add required global parameters
-        self.global_params.add_parameters([
-            TorchParameter(vco_ratio, "vco_ratio", ParameterRange(0.0, 1.0))
-        ])
-
-        # Register all modules as children
-        self.add_synth_modules({
-            'pitch_adsr': pitch_adsr,
-            'amp_adsr': amp_adsr,
-            'vco_1': vco_1,
-            'vco_2': vco_2,
-            'noise': noise,
-            'vca': vca
-        })
-
-    def forward(self) -> T:
-        # The convention for triggering a note event is that it has
-        # the same note_on_duration for both ADSRs.
-        note_on_duration = self.note_on_duration
-        pitch_envelope = self.pitch_adsr(note_on_duration)
-        amp_envelope = self.amp_adsr(note_on_duration)
-
-        vco_1_out = self.vco_1(pitch_envelope)
-        vco_2_out = self.vco_2(pitch_envelope)
-
-        audio_out = util.crossfade(
-            vco_1_out,
-            vco_2_out,
-            self.global_params.p("vco_ratio")
-        )
-
-        audio_out = self.noise(audio_out)
-
-        return self.vca(amp_envelope, audio_out)
-
-
-class FIRLowPass(TorchSynthModule):
-    """
-    A finite impulse response low-pass filter. Uses convolution with a windowed
-    sinc function.
-
-    Parameters
-    ----------
-
-    cutoff (float)      :   cutoff frequency of low-pass in Hz, must be between 5 and
-                            half the sampling rate. Defaults to 1000Hz.
-    filter_length (int) :   The length of the filter in samples. A longer filter will
-                            result in a steeper filter cutoff. Should be greater than 4.
-                            Defaults to 512 samples.
-    sample_rate (int)   :   Sampling rate to run processing at.
-    """
-
-    def __init__(
-        self,
-        cutoff: float = 1000.0,
-        filter_length: int = 512,
-        sample_rate: int = SAMPLE_RATE,
-    ):
-        super().__init__(sample_rate=sample_rate)
-        self.add_parameters(
-            [
-                TorchParameter(
-                    value=cutoff,
-                    parameter_name="cutoff",
-                    parameter_range=ParameterRange(5.0, sample_rate / 2.0, "log")
-                ),
-                TorchParameter(
-                    value=filter_length,
-                    parameter_name="length",
-                    parameter_range=ParameterRange(4.0, 4096.)
-                )
-            ]
-        )
-
-    def _forward(self, audio_in: T) -> T:
-        """
-        Filter audio samples
-        TODO: Cutoff frequency modulation, if there is an efficient way to do it
-
-        Parameters
-        ----------
-
-        audio (T)  :   audio samples to filter
-        """
-
-        impulse = self.windowed_sinc(self.p("cutoff"), self.p("length"))
-        impulse = impulse.view(1, 1, impulse.size()[0])
-        audio_resized = audio_in.view(1, 1, audio_in.size()[0])
-        y = nn.functional.conv1d(
-            audio_resized,
-            impulse,
-            padding=int(self.p("length") / 2)
-        )
-        return y[0][0]
-
-    def windowed_sinc(self, cutoff: T, length: T) -> T:
-        """
-        Calculates the impulse response for FIR low-pass filter using the
-        windowed sinc function method. Updated to allow for a fractional filter length.
-
-        Parameters
-        ----------
-
-        cutoff (T)      :   Low-pass cutoff frequency in Hz. Must be between 0 and
-                                half the sampling rate.
-        length (T) :   Length of the filter impulse response to create.
-        """
-
-        # Normalized frequency
-        omega = 2 * torch.pi * cutoff / self.sample_rate
-
-        # Create a sinc function
-        num_samples = torch.ceil(length)
-        half_length = (length - 1.) / 2.
-        t = torch.arange(num_samples.detach(), device=length.device)
-        ir = util.sinc((t - half_length) * omega)
-
-        return ir * util.blackman(length)
-
-
-class TorchMovingAverage(TorchSynthModule):
-    """
-    A finite impulse response moving average filter.
-
-    Parameters
-    ----------
-
-    filter_length (int) :   Length of filter and number of samples to take average over.
-                            Must be greater than 0. Defaults to 32.
-    sample_rate (int)   :   Sampling rate to run processing at.
-    """
-
-    def __init__(self, filter_length: int = 32, sample_rate: int = SAMPLE_RATE):
-        super().__init__(sample_rate=sample_rate)
-        self.add_parameters(
-            [
-                TorchParameter(
-                    value=filter_length,
-                    parameter_name="length",
-                    parameter_range=ParameterRange(1.0, 4096.0),
-                )
-            ]
-        )
-
-    def _forward(self, audio_in: T) -> T:
-        """
-        Filter audio samples
-
-        Parameters
-        ----------
-
-        audio (T)  :   audio samples to filter
-        """
-        length = self.p("length")
-        impulse = torch.ones((1, 1, int(length)), device=length.device) / length
-
-        # For non-integer impulse lengths
-        if torch.sum(impulse) < 1.0:
-            additional = torch.ones(1, 1, 1, device=length.device)
-            additional *= (1.0 - torch.sum(impulse))
-            impulse = torch.cat((impulse, additional), dim=2)
-
-        audio_resized = audio_in.view(1, 1, audio_in.size()[0])
-        y = nn.functional.conv1d(
-            audio_resized,
-            impulse,
-            padding=int(impulse.size()[0] / 2)
-        )
-        return y[0][0]
-
-
-class TorchSVF(TorchSynthModule):
-    """
-    A State Variable Filter that can do low-pass, high-pass, band-pass, and
-    band-reject filtering. Allows modulation of the cutoff frequency and an
-    adjustable resonance parameter. Can self-oscillate to make a sinusoid
-    oscillator.
-
-    Parameters
-    ----------
-
-    mode (str)              :   filter type, one of LPF, HPF, BPF, or BSF
-    cutoff (float)          :   cutoff frequency in Hz must be between 5 and
-                                half the sample rate. Defaults to 1000Hz.
-    resonance (float)       :   filter resonance, or "Quality Factor". Higher
-                                values cause the filter to resonate more. Must
-                                be greater than 0.5. Defaults to 0.707.
-    mod_depth (float)       :   Amount of modulation to apply to the cutoff from
-                                the control input during processing. Can be negative
-                                or positive in Hertz. Defaults to zero.
-    """
-
-    def __init__(
-            self,
-            mode: str,
-            cutoff: float = 1000.0,
-            resonance: float = 0.707,
-            mod_depth: float = 0.0,
-            **kwargs
-    ):
-        super().__init__(**kwargs)
-
-        # Set the filter type
-        self.mode = mode.lower()
-        assert mode in ['lpf', 'hpf', 'bpf', 'bsf']
-
-        nyquist = self.sample_rate / 2.0
-        self.add_parameters(
-            [
-                TorchParameter(
-                    value=cutoff,
-                    parameter_range=ParameterRange(5.0, nyquist, "log"),
-                    parameter_name="cutoff"
-                ),
-                TorchParameter(
-                    value=resonance,
-                    parameter_range=ParameterRange(0.01, 1000.0, "log"),
-                    parameter_name="resonance"
-                ),
-                TorchParameter(
-                    value=mod_depth,
-                    parameter_range=ParameterRange(-nyquist, nyquist, "log"),
-                    parameter_name="mod_depth"
-                )
-            ]
-        )
-
-    def _forward(
-        self,
-        audio_in: T,
-        control_in: T = None,
-    ) -> T:
-        """
-        Process audio samples and return filtered results.
-
-        Parameters
-        ----------
-
-        audio (torch.tensor)          :   Audio samples to filter
-        cutoff_mod (torch.tensor)     :   Control signal used to modulate the filter
-                                        cutoff. Values must be in range [0,1]
-        """
-
-        h0 = 0.0
-        h1 = 0.0
-        y = torch.zeros_like(audio_in, device=audio_in.device)
-
-        if control_in is None:
-            control_in = torch.zeros_like(audio_in, device=audio_in.device)
-        else:
-            assert control_in.size() == audio_in.size()
-
-        cutoff = self.p("cutoff")
-        mod_depth = self.p("mod_depth")
-        res_coefficient = 1.0 / self.p("resonance")
-
-        # Processing loop
-        for i in range(len(audio_in)):
-            # If there is a cutoff modulation envelope, update coefficients
-            cutoff_val = cutoff + control_in[i] * mod_depth
-            coeff0, coeff1, rho = TorchSVF.svf_coefficients(
-                cutoff_val,
-                res_coefficient,
-                self.sample_rate
-            )
-
-            # Calculate each of the filter components
-            hpf = coeff0 * (audio_in[i] - rho * h0 - h1)
-            bpf = coeff1 * hpf + h0
-            lpf = coeff1 * bpf + h1
-
-            # Feedback samples
-            h0 = coeff1 * hpf + bpf
-            h1 = coeff1 * bpf + lpf
-
-            if self.mode == "lpf":
-                y[i] = lpf
-            elif self.mode == "bpf":
-                y[i] = bpf
-            elif self.mode == "bsf":
-                y[i] = hpf + lpf
-            else:
-                y[i] = hpf
-
-        return y
-
-    @staticmethod
-    def svf_coefficients(cutoff, res_coefficient, sample_rate):
-        """
-        Calculates the filter coefficients for SVF.
-
-        Parameters
-        ----------
-        cutoff (T)  :   Filter cutoff frequency in Hz.
-        resonance (T) : Filter resonance
-        sample_rate (T) : Sample rate to process at
-        """
-
-        g = torch.tan(torch.pi * cutoff / sample_rate)
-        coeff0 = 1.0 / (1.0 + res_coefficient * g + g * g)
-        rho = res_coefficient + g
-
-        return coeff0, g, rho
-
-
-class TorchLowPassSVF(TorchSVF):
-    """
-    IIR Low-pass using SVF architecture
-
-    Parameters
-    ----------
-    kwargs: see TorchSVF
-    """
-
-    def __init__(self, **kwargs):
-        super().__init__('lpf', **kwargs)
-
-
-class TorchHighPassSVF(TorchSVF):
-    """
-    IIR High-pass using SVF architecture
-
-    Parameters
-    ----------
-    kwargs: see TorchSVF
-    """
-
-    def __init__(self, **kwargs):
-        super().__init__('hpf', **kwargs)
-
-
-class TorchBandPassSVF(TorchSVF):
-    """
-    IIR Band-pass using SVF architecture
-
-    Parameters
-    ----------
-    kwargs: see TorchSVF
-    """
-
-    def __init__(self, **kwargs):
-        super().__init__('bpf', **kwargs)
-
-
-class TorchBandStopSVF(TorchSVF):
-    """
-    IIR Band-stop using SVF architecture
-
-    Parameters
-    ----------
-    kwargs: see TorchSVF
-    """
-
-    def __init__(self, **kwargs):
-        super().__init__('bsf', **kwargs)
+#class TorchDrum(TorchSynth):
+#    """
+#    A package of modules that makes one drum hit.
+#    """
+#
+#    def __init__(
+#        self,
+#        note_on_duration: float,
+#        vco_ratio: float = 0.5,
+#        pitch_adsr: TorchADSR = TorchADSR(),
+#        amp_adsr: TorchADSR = TorchADSR(),
+#        vco_1: TorchVCO = TorchSineVCO(),
+#        vco_2: TorchVCO = TorchSquareSawVCO(),
+#        noise: TorchNoise = TorchNoise(),
+#        vca: TorchVCA = TorchVCA(),
+#        **kwargs,
+#    ):
+#        super().__init__(**kwargs)
+#        assert note_on_duration >= 0
+#
+#        # We assume that sustain duration is a hyper-parameter,
+#        # with the mindset that if you are trying to learn to
+#        # synthesize a sound, you won't be adjusting the note_on_duration.
+#        # However, this is easily changed if desired.
+#        self.note_on_duration = T(note_on_duration)
+#
+#        # Add required global parameters
+#        self.global_params.add_parameters(
+#            [TorchParameter(vco_ratio, "vco_ratio", ParameterRange(0.0, 1.0))]
+#        )
+#
+#        # Register all modules as children
+#        self.add_synth_modules(
+#            {
+#                "pitch_adsr": pitch_adsr,
+#                "amp_adsr": amp_adsr,
+#                "vco_1": vco_1,
+#                "vco_2": vco_2,
+#                "noise": noise,
+#                "vca": vca,
+#            }
+#        )
+#
+#    def forward(self) -> T:
+#        # The convention for triggering a note event is that it has
+#        # the same note_on_duration for both ADSRs.
+#        note_on_duration = self.note_on_duration
+#        pitch_envelope = self.pitch_adsr(note_on_duration)
+#        amp_envelope = self.amp_adsr(note_on_duration)
+#
+#        vco_1_out = self.vco_1(pitch_envelope)
+#        vco_2_out = self.vco_2(pitch_envelope)
+#
+#        audio_out = util.crossfade(
+#            vco_1_out, vco_2_out, self.global_params.p("vco_ratio")
+#        )
+#
+#        audio_out = self.noise(audio_out)
+#
+#        return self.vca(amp_envelope, audio_out)
